+++ examples/impex.py
@@ -26,7 +26,7 @@ def export_entities_with(impex: Spark.ImpEx):
 
 def import_entities_with(impex: Spark.ImpEx):
     destination = {'source': 'my-folder/my-service', 'target': 'this-folder/my-service', 'upgrade': 'patch'}
-    response = impex.run_import(destination, file=open('exported.zip', 'rb'), max_retries=7, retry_interval=3)
+    response = impex.import_(destination, file=open('exported.zip', 'rb'), max_retries=7, retry_interval=3)
     print(response.data)